refactor: deprecate blobtestutil
