+++ app/estimate_square_size_test.go
@@ -43,8 +43,8 @@ func Test_estimateSquareSize(t *testing.T) {
 	signer := types.GenerateKeyringSigner(t, testEstimateKey)
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			txs := generateManyRawWirePFB(t, encConf.TxConfig, signer, tt.wPFBCount, tt.messgeSize)
-			txs = append(txs, generateManyRawSendTxs(t, encConf.TxConfig, signer, tt.normalTxs)...)
+			txs := GenerateManyRawWirePFB(t, encConf.TxConfig, signer, tt.wPFBCount, tt.messgeSize)
+			txs = append(txs, GenerateManyRawSendTxs(t, encConf.TxConfig, signer, tt.normalTxs)...)
 			parsedTxs := parseTxs(encConf.TxConfig, txs)
 			squareSize, totalSharesUsed := estimateSquareSize(parsedTxs, core.EvidenceList{})
 			assert.Equal(t, tt.expectedSize, squareSize)
@@ -75,8 +75,8 @@ func Test_estimateSquareSize(t *testing.T) {
 func Test_pruning(t *testing.T) {
 	encConf := encoding.MakeConfig(ModuleEncodingRegisters...)
 	signer := types.GenerateKeyringSigner(t, testEstimateKey)
-	txs := generateManyRawSendTxs(t, encConf.TxConfig, signer, 10)
-	txs = append(txs, generateManyRawWirePFB(t, encConf.TxConfig, signer, 10, 1000)...)
+	txs := GenerateManyRawSendTxs(t, encConf.TxConfig, signer, 10)
+	txs = append(txs, GenerateManyRawWirePFB(t, encConf.TxConfig, signer, 10, 1000)...)
 	parsedTxs := parseTxs(encConf.TxConfig, txs)
 	ss, total := estimateSquareSize(parsedTxs, core.EvidenceList{})
 	nextLowestSS := ss / 2
@@ -170,8 +170,8 @@ func Test_calculateCompactShareCount(t *testing.T) {
 	signer := types.GenerateKeyringSigner(t, testEstimateKey)
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			txs := generateManyRawWirePFB(t, encConf.TxConfig, signer, tt.wPFBCount, tt.messgeSize)
-			txs = append(txs, generateManyRawSendTxs(t, encConf.TxConfig, signer, tt.normalTxs)...)
+			txs := GenerateManyRawWirePFB(t, encConf.TxConfig, signer, tt.wPFBCount, tt.messgeSize)
+			txs = append(txs, GenerateManyRawSendTxs(t, encConf.TxConfig, signer, tt.normalTxs)...)
 
 			parsedTxs := parseTxs(encConf.TxConfig, txs)
 			squareSize, totalSharesUsed := estimateSquareSize(parsedTxs, core.EvidenceList{})
