+++ app/test/process_proposal_test.go
@@ -15,7 +15,6 @@ import (
 	"github.com/celestiaorg/celestia-app/app/encoding"
 	"github.com/celestiaorg/celestia-app/pkg/appconsts"
 	"github.com/celestiaorg/celestia-app/testutil"
-	paytestutil "github.com/celestiaorg/celestia-app/testutil/blob"
 	"github.com/celestiaorg/celestia-app/x/blob/types"
 	"github.com/celestiaorg/nmt/namespace"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -29,7 +28,7 @@ func TestBlobInclusionCheck(t *testing.T) {
 	// block with all blobs included
 	validData := func() *core.Data {
 		return &core.Data{
-			Txs: paytestutil.GenerateManyRawWirePFB(t, encConf.TxConfig, signer, 4, 1000),
+			Txs: app.GenerateManyRawWirePFB(t, encConf.TxConfig, signer, 4, 1000),
 		}
 	}
